+++ src/main/java/com/pbs/ams/web/model/AmsTradeInstruct.java
@@ -8,7 +8,7 @@ public class AmsTradeInstruct implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer instructId;
+    private Long instructId;
 
     /**
      * 交易类别(股票1/期货2/期权3)
@@ -29,7 +29,7 @@ public class AmsTradeInstruct implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer instructUserId;
+    private Long instructUserId;
 
     /**
      * 备注
@@ -92,15 +92,15 @@ public class AmsTradeInstruct implements Serializable {
      *
      * @mbg.generated
      */
-    private Integer operatorId;
+    private Long operatorId;
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getInstructId() {
+    public Long getInstructId() {
         return instructId;
     }
 
-    public void setInstructId(Integer instructId) {
+    public void setInstructId(Long instructId) {
         this.instructId = instructId;
     }
 
@@ -120,11 +120,11 @@ public class AmsTradeInstruct implements Serializable {
         this.tradeOrderType = tradeOrderType;
     }
 
-    public Integer getInstructUserId() {
+    public Long getInstructUserId() {
         return instructUserId;
     }
 
-    public void setInstructUserId(Integer instructUserId) {
+    public void setInstructUserId(Long instructUserId) {
         this.instructUserId = instructUserId;
     }
 
@@ -192,11 +192,11 @@ public class AmsTradeInstruct implements Serializable {
         this.updateTime = updateTime;
     }
 
-    public Integer getOperatorId() {
+    public Long getOperatorId() {
         return operatorId;
     }
 
-    public void setOperatorId(Integer operatorId) {
+    public void setOperatorId(Long operatorId) {
         this.operatorId = operatorId;
     }
 }
